+++ gomint-server/src/main/java/io/gomint/server/world/block/ChorusPlant.java
@@ -13,8 +13,8 @@ import io.gomint.server.registry.RegisterInfo;
 public class ChorusPlant extends Block implements io.gomint.world.block.BlockChorusPlant {
 
     @Override
-    public byte getBlockId() {
-        return (byte) 240;
+    public int getBlockId() {
+        return 240;
     }
 
     @Override
