+++ gomint-server/src/main/java/io/gomint/server/world/block/CommandBlock.java
@@ -15,8 +15,8 @@ import io.gomint.server.registry.RegisterInfo;
 public class CommandBlock extends ContainerBlock implements io.gomint.world.block.BlockCommandBlock {
 
     @Override
-    public byte getBlockId() {
-        return (byte) 137;
+    public int getBlockId() {
+        return 137;
     }
 
     @Override
