+++ gomint-server/src/main/java/io/gomint/server/world/block/DarkOakFenceGate.java
@@ -12,8 +12,8 @@ import io.gomint.server.registry.RegisterInfo;
 public class DarkOakFenceGate extends Block implements io.gomint.world.block.BlockDarkOakFenceGate {
 
     @Override
-    public byte getBlockId() {
-        return (byte) 186;
+    public int getBlockId() {
+        return 186;
     }
 
     @Override
