+++ gomint-server/src/main/java/io/gomint/server/world/block/NetherQuartzOre.java
@@ -14,8 +14,8 @@ import io.gomint.server.registry.RegisterInfo;
 public class NetherQuartzOre extends Block implements io.gomint.world.block.BlockNetherQuartzOre {
 
     @Override
-    public byte getBlockId() {
-        return (byte) 153;
+    public int getBlockId() {
+        return 153;
     }
 
     @Override
