+++ gomint-server/src/main/java/io/gomint/server/world/block/NetherWartBlock.java
@@ -20,8 +20,8 @@ import io.gomint.world.block.BlockType;
 public class NetherWartBlock extends Block implements BlockNetherWartBlock {
 
     @Override
-    public byte getBlockId() {
-        return (byte) 214;
+    public int getBlockId() {
+        return 214;
     }
 
     @Override
